infra/gcp/function/requirements.txt
flask
functions-framework
langchain
langchain-community
langchain-google-vertexai
#langchain-google-community
google-cloud-aiplatform==1.95.1
google-auth
python-dotenv
google-cloud-discoveryengine
langchain_google_genai
requests
types-requests